Extract pod log helpers into internal/podlogs
Move the pod selector resolution, container log fetching, byte-bounded
tailing and chronological stream merging out of the status page backend
into a shared internal/podlogs package, so the MCP server can reuse the
same log processing logic instead of duplicating it.

The web handlers keep their behaviour and now call the exported helpers.
