+++ internal/web/workload_logs.go
@@ -7,9 +7,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io"
 	"net/http"
-	"sort"
 	"strconv"
 	"strings"
 	"sync"
@@ -21,6 +19,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
+	"github.com/controlplaneio-fluxcd/flux-operator/internal/podlogs"
 	"github.com/controlplaneio-fluxcd/flux-operator/internal/web/user"
 )
 
@@ -100,108 +99,6 @@ type WorkloadLogsResponse struct {
 	Forbidden int `json:"forbidden,omitempty"`
 }
 
-// trimPartialLogLine drops a trailing partial log line from the payload.
-//
-// Container runtimes newline-terminate every emitted log line, so a payload
-// that does not end with a newline has been truncated mid-line because the
-// container was writing the line when the logs were read (common while
-// following). Returning that fragment as the latest entry shows only a few
-// characters, so it is dropped; the complete line reappears on the next fetch.
-// A payload with no newline at all is returned unchanged so a single short line
-// is not lost.
-func trimPartialLogLine(logs string) string {
-	if logs == "" || logs[len(logs)-1] == '\n' {
-		return logs
-	}
-	if i := strings.LastIndexByte(logs, '\n'); i >= 0 {
-		return logs[:i+1]
-	}
-	return logs
-}
-
-// tailLogBytes reads r to EOF but retains only the most recent limit bytes, so
-// the response stays small while returning the newest log lines.
-//
-// The kubelet's PodLogOptions.LimitBytes keeps the *oldest* bytes of the range
-// it reads, which for a tail window drops the most recent lines. Reading the
-// whole window (bounded by TailLines) and keeping the trailing bytes here
-// returns the newest lines instead. Memory use is bounded to ~limit regardless
-// of the stream length.
-//
-// The returned bool is partialFirst: true when the retained slice begins
-// mid-line (the byte dropped just before it was not a newline), so the caller
-// should drop that leading fragment. A cut that lands exactly on a line boundary
-// keeps a complete first line and reports false.
-func tailLogBytes(r io.Reader, limit int) ([]byte, bool, error) {
-	const chunkSize = 32 * 1024
-	buf := make([]byte, 0, limit+chunkSize)
-	chunk := make([]byte, chunkSize)
-	partialFirst := false
-	for {
-		n, err := r.Read(chunk)
-		if n > 0 {
-			buf = append(buf, chunk[:n]...)
-			if len(buf) > limit {
-				// The byte dropped just before the retained window decides
-				// whether the first kept line is complete (preceded by '\n') or
-				// a mid-line fragment.
-				partialFirst = buf[len(buf)-limit-1] != '\n'
-				// Shift the trailing limit bytes (the newest output) to the
-				// front in place so the backing array, and thus memory, stays
-				// bounded.
-				copy(buf, buf[len(buf)-limit:])
-				buf = buf[:limit]
-			}
-		}
-		if err == io.EOF {
-			return buf, partialFirst, nil
-		}
-		if err != nil {
-			return nil, partialFirst, err
-		}
-	}
-}
-
-// trimPartialFirstLine drops a leading partial log line from the payload. It is
-// applied only when the payload was byte-truncated from the front (see
-// tailLogBytes), where the first line is a mid-line fragment. A payload with no
-// newline is returned unchanged so a single short line is not lost.
-func trimPartialFirstLine(logs string) string {
-	if _, rest, found := strings.Cut(logs, "\n"); found {
-		return rest
-	}
-	return logs
-}
-
-// fetchContainerLog streams the logs of a single container with the given
-// options and returns the payload trimmed of any partial first/last line,
-// keeping only the newest byteLimit bytes. It is shared by the single-stream and
-// fan-out paths of the handler; the fan-out passes a smaller per-stream limit so
-// many concurrent streams stay within the overall byte budget.
-func fetchContainerLog(ctx context.Context, clientset kubernetes.Interface, namespace, name string, opts *corev1.PodLogOptions, byteLimit int) (string, error) {
-	stream, err := clientset.CoreV1().Pods(namespace).GetLogs(name, opts).Stream(ctx)
-	if err != nil {
-		return "", err
-	}
-	defer stream.Close()
-
-	// Read the stream, keeping only the newest byteLimit bytes so the payload
-	// stays bounded while returning the most recent lines.
-	data, partialFirst, err := tailLogBytes(stream, byteLimit)
-	if err != nil {
-		return "", err
-	}
-
-	// Drop the partial first line only when the byte cap cut mid-line (a cut on
-	// a line boundary keeps a complete first line), and the partial last line
-	// when the container was mid-write.
-	logs := trimPartialLogLine(string(data))
-	if partialFirst {
-		logs = trimPartialFirstLine(logs)
-	}
-	return logs, nil
-}
-
 // writeLogStreamError maps a GetLogs stream error to an HTTP response: 403 for a
 // forbidden user, 404 for a missing pod, and 500 otherwise (logged). An empty
 // name produces a workload-scoped message for the all-pods fan-out, where the
@@ -248,144 +145,6 @@ func writeWorkloadLogsResponse(w http.ResponseWriter, resp WorkloadLogsResponse)
 	}
 }
 
-// logTarget is one (pod, container) stream to read in the fan-out. An empty
-// container lets the kubelet pick the pod's default container.
-type logTarget struct {
-	pod       string
-	container string
-}
-
-// logStream is one source blob in the fan-out: the raw timestamped log payload
-// of a single (pod, container) target plus the pod and container it came from, so
-// the merge can tag each line with its origin in the all-pods/all-containers view.
-type logStream struct {
-	pod       string
-	container string
-	blob      string
-}
-
-// logEntry is one record used when interleaving multiple streams: a line
-// carrying a leading RFC3339 timestamp plus any immediately following
-// continuation lines that lack their own timestamp (e.g. stack-trace frames).
-// The parsed timestamp is the sort key; text retains the original line(s),
-// including the timestamp prefix, since the frontend strips it for display. pod
-// and container record the origin so the merge can prefix the line in tagged mode.
-type logEntry struct {
-	ts        time.Time
-	text      string
-	pod       string
-	container string
-}
-
-// parseLogTimestamp parses the leading RFC3339 timestamp token kubelet prepends
-// to each line when PodLogOptions.Timestamps is set. It reports false when the
-// line does not start with such a token (e.g. a stack-trace continuation).
-func parseLogTimestamp(line string) (time.Time, bool) {
-	tsStr, _, found := strings.Cut(line, " ")
-	if !found {
-		return time.Time{}, false
-	}
-	t, err := time.Parse(time.RFC3339Nano, tsStr)
-	if err != nil {
-		return time.Time{}, false
-	}
-	return t, true
-}
-
-// parseLogEntries splits a timestamped log payload into entries, stamping each
-// with pod and container. Each entry begins at a line whose first token parses as
-// an RFC3339 timestamp and absorbs subsequent lines without such a prefix, so
-// multi-line records (stack traces) stay intact and ordered. A continuation line
-// with no preceding entry becomes its own zero-timestamped entry, which sorts to
-// the front.
-func parseLogEntries(blob, pod, container string) []logEntry {
-	var entries []logEntry
-	for line := range strings.SplitSeq(blob, "\n") {
-		if line == "" {
-			continue
-		}
-		if ts, ok := parseLogTimestamp(line); ok {
-			entries = append(entries, logEntry{ts: ts, text: line, pod: pod, container: container})
-			continue
-		}
-		if n := len(entries); n > 0 {
-			entries[n-1].text += "\n" + line
-		} else {
-			entries = append(entries, logEntry{text: line, pod: pod, container: container})
-		}
-	}
-	return entries
-}
-
-// mergeLogStreams interleaves the timestamped payloads of multiple streams into
-// a single chronological stream. Entries are stable-sorted by their timestamp
-// (so records logged at the same instant keep a deterministic order), capped to
-// the newest tailLines entries, and finally to the newest maxLogBytes bytes on a
-// line boundary so the payload stays bounded.
-//
-// Each entry that carries a timestamp is prefixed with its origin tags, in
-// pod-then-container order, for whichever dimensions fanned out: tagPod writes the
-// pod, tagContainer the container ("<pod> <container> <timestamp> <message>", or
-// just one token when only one dimension is tagged), so the client can attribute
-// the line and scope multi-line folding. Zero-timestamped orphan entries are left
-// untagged so every tagged line is uniformly "<tags...> <timestamp> ..." and
-// parses unambiguously. With both flags false the output is byte-for-byte the
-// original single-pod merge.
-func mergeLogStreams(streams []logStream, tailLines int, tagPod, tagContainer bool) string {
-	var entries []logEntry
-	for _, s := range streams {
-		entries = append(entries, parseLogEntries(s.blob, s.pod, s.container)...)
-	}
-	sort.SliceStable(entries, func(i, j int) bool {
-		return entries[i].ts.Before(entries[j].ts)
-	})
-	if tailLines > 0 && len(entries) > tailLines {
-		entries = entries[len(entries)-tailLines:]
-	}
-	var sb strings.Builder
-	for _, e := range entries {
-		if !e.ts.IsZero() {
-			if tagPod {
-				sb.WriteString(e.pod)
-				sb.WriteByte(' ')
-			}
-			if tagContainer {
-				sb.WriteString(e.container)
-				sb.WriteByte(' ')
-			}
-		}
-		sb.WriteString(e.text)
-		sb.WriteByte('\n')
-	}
-	return capLogBytes(sb.String(), int(maxLogBytes))
-}
-
-// dedupeNames removes duplicate names (keeping first occurrence) and caps the
-// result to limit, so a request naming the same pod or container twice does not
-// stream it twice and a request naming many cannot fan out without bound. Order
-// is preserved so the merge input stays deterministic. It reports whether any
-// name was dropped by the cap (not by de-duplication).
-func dedupeNames(names []string, limit int) ([]string, bool) {
-	if len(names) == 0 {
-		return names, false
-	}
-	seen := make(map[string]struct{}, len(names))
-	out := make([]string, 0, len(names))
-	truncated := false
-	for _, n := range names {
-		if _, ok := seen[n]; ok {
-			continue
-		}
-		seen[n] = struct{}{}
-		if len(out) == limit {
-			truncated = true
-			break
-		}
-		out = append(out, n)
-	}
-	return out, truncated
-}
-
 // logFanOut is the outcome of a per-target log fan-out: the successful streams
 // (tagged with their pod), the set of distinct pods that produced logs, the
 // number of distinct pods skipped because the user is forbidden to read them, and
@@ -394,7 +153,7 @@ func dedupeNames(names []string, limit int) ([]string, bool) {
 // deleted, or forbidden) is skipped as long as another succeeded, so firstErr is
 // only meaningful when no stream was collected.
 type logFanOut struct {
-	streams     []logStream
+	streams     []podlogs.LogStream
 	streamedSet map[string]struct{}
 	forbidden   int
 	firstErr    error
@@ -405,7 +164,7 @@ type logFanOut struct {
 // pods forbidden, and the first error. targets, logs, and errs are index-aligned.
 // Forbidden is counted per pod (not per target) so a forbidden pod with several
 // containers is reported once, matching the pod-based Total/Streamed fields.
-func collectLogStreams(targets []logTarget, logs []string, errs []error) logFanOut {
+func collectLogStreams(targets []podlogs.LogTarget, logs []string, errs []error) logFanOut {
 	out := logFanOut{streamedSet: make(map[string]struct{})}
 	forbiddenPods := make(map[string]struct{})
 	for i := range targets {
@@ -414,35 +173,15 @@ func collectLogStreams(targets []logTarget, logs []string, errs []error) logFanO
 				out.firstErr = errs[i]
 			}
 			if errors.IsForbidden(errs[i]) {
-				forbiddenPods[targets[i].pod] = struct{}{}
+				forbiddenPods[targets[i].Pod] = struct{}{}
 			}
 			continue
 		}
-		out.streams = append(out.streams, logStream{pod: targets[i].pod, container: targets[i].container, blob: logs[i]})
-		out.streamedSet[targets[i].pod] = struct{}{}
+		out.streams = append(out.streams, podlogs.LogStream{Pod: targets[i].Pod, Container: targets[i].Container, Blob: logs[i]})
+		out.streamedSet[targets[i].Pod] = struct{}{}
 	}
 	out.forbidden = len(forbiddenPods)
 	return out
-}
-
-// capLogBytes keeps only the newest limit bytes of a newline-terminated log
-// payload, trimming any partial leading line so the result starts on a line
-// boundary.
-func capLogBytes(logs string, limit int) string {
-	if len(logs) <= limit {
-		return logs
-	}
-	cut := len(logs) - limit
-	// When the cut lands exactly on a line boundary (the preceding byte is a
-	// newline), the window already starts with a complete line and must not be
-	// trimmed further; otherwise drop the leading partial fragment.
-	if logs[cut-1] == '\n' {
-		return logs[cut:]
-	}
-	if i := strings.IndexByte(logs[cut:], '\n'); i >= 0 {
-		return logs[cut+i+1:]
-	}
-	return logs[cut:]
 }
 
 // WorkloadLogsHandler handles GET /api/v1/workload/logs requests and returns the
@@ -579,9 +318,9 @@ func (h *Handler) WorkloadLogsHandler(w http.ResponseWriter, req *http.Request) 
 	}
 	// total is the number of distinct pods the client asked for, measured before
 	// the cap so a truncated all-pods view can be reported as partial.
-	deduped, _ := dedupeNames(pods, len(pods))
+	deduped, _ := podlogs.DedupeNames(pods, len(pods))
 	total := len(deduped)
-	pods, podsTruncated := dedupeNames(pods, maxLogPods)
+	pods, podsTruncated := podlogs.DedupeNames(pods, maxLogPods)
 
 	// Collect the requested container names. An absent container lets the kubelet
 	// pick the pod's default container.
@@ -591,7 +330,7 @@ func (h *Handler) WorkloadLogsHandler(w http.ResponseWriter, req *http.Request) 
 			containers = append(containers, c)
 		}
 	}
-	containers, containersTruncated := dedupeNames(containers, maxLogContainers)
+	containers, containersTruncated := podlogs.DedupeNames(containers, maxLogContainers)
 
 	// Tagging is decided from the client's request, before the stream cap, so the
 	// client and server agree on the wire format regardless of which streams
@@ -604,7 +343,7 @@ func (h *Handler) WorkloadLogsHandler(w http.ResponseWriter, req *http.Request) 
 	// Build the (pod, container) targets as the product of the two dimensions,
 	// capped to maxLogStreams overall. An empty container list yields one target
 	// per pod (kubelet default container).
-	targets, streamsTruncated := buildLogTargets(pods, containers, maxLogStreams)
+	targets, streamsTruncated := podlogs.BuildLogTargets(pods, containers, maxLogStreams)
 	truncated := podsTruncated || containersTruncated || streamsTruncated
 
 	// TailLines is always set so the kubelet returns the newest lines (and bounds
@@ -626,12 +365,12 @@ func (h *Handler) WorkloadLogsHandler(w http.ResponseWriter, req *http.Request) 
 	// (maxLogStreams is 50). Were a future cap to collapse it to one target, the
 	// response would report ContainerTagged=false, so client and server still agree.
 	if !tagPod && len(targets) <= 1 {
-		t := logTarget{pod: name}
+		t := podlogs.LogTarget{Pod: name}
 		if len(targets) == 1 {
 			t = targets[0]
 		}
 		opts := &corev1.PodLogOptions{
-			Container:  t.container,
+			Container:  t.Container,
 			TailLines:  &tailLines64,
 			Previous:   previous,
 			Timestamps: true,
@@ -639,12 +378,12 @@ func (h *Handler) WorkloadLogsHandler(w http.ResponseWriter, req *http.Request) 
 		if sinceTime != nil {
 			opts.SinceTime = sinceTime
 		}
-		logs, err := fetchContainerLog(ctx, clientset, namespace, t.pod, opts, int(maxLogBytes))
+		logs, err := podlogs.FetchContainerLog(ctx, clientset, namespace, t.Pod, opts, int(maxLogBytes))
 		if err != nil {
-			writeLogStreamError(ctx, w, err, namespace, t.pod, t.container)
+			writeLogStreamError(ctx, w, err, namespace, t.Pod, t.Container)
 			return
 		}
-		writeWorkloadLogs(w, t.pod, t.container, logs)
+		writeWorkloadLogs(w, t.Pod, t.Container, logs)
 		return
 	}
 
@@ -662,24 +401,24 @@ func (h *Handler) WorkloadLogsHandler(w http.ResponseWriter, req *http.Request) 
 	var wg sync.WaitGroup
 	for i, t := range targets {
 		wg.Add(1)
-		go func(i int, t logTarget) {
+		go func(i int, t podlogs.LogTarget) {
 			defer wg.Done()
 			sem <- struct{}{}
 			defer func() { <-sem }()
 			opts := &corev1.PodLogOptions{
-				Container:  t.container,
+				Container:  t.Container,
 				TailLines:  &tailLines64,
 				Timestamps: true,
 			}
 			// Follow cursor: a per-pod `since` for the all-pods view, falling back
 			// to the single global sinceTime (used by a single-pod, multi-container
 			// follow, which routes here yet sends one cursor).
-			if st := sinceByPod[t.pod]; st != nil {
+			if st := sinceByPod[t.Pod]; st != nil {
 				opts.SinceTime = st
 			} else if sinceTime != nil {
 				opts.SinceTime = sinceTime
 			}
-			logsByTarget[i], errsByTarget[i] = fetchContainerLog(ctx, clientset, namespace, t.pod, opts, perStreamBytes)
+			logsByTarget[i], errsByTarget[i] = podlogs.FetchContainerLog(ctx, clientset, namespace, t.Pod, opts, perStreamBytes)
 		}(i, t)
 	}
 	wg.Wait()
@@ -711,7 +450,7 @@ func buildWorkloadLogsResponse(pods, containers []string, fanOut logFanOut, tota
 	return WorkloadLogsResponse{
 		Pod:             strings.Join(pods, ","),
 		Container:       strings.Join(containers, ","),
-		Logs:            mergeLogStreams(fanOut.streams, tailLines, tagPod, tagContainer),
+		Logs:            podlogs.MergeLogStreams(fanOut.streams, tailLines, tagPod, tagContainer, int(maxLogBytes)),
 		Tagged:          tagPod,
 		ContainerTagged: tagContainer,
 		Total:           total,
@@ -719,28 +458,4 @@ func buildWorkloadLogsResponse(pods, containers []string, fanOut logFanOut, tota
 		Partial:         streamed < total || truncated,
 		Forbidden:       fanOut.forbidden,
 	}
-}
-
-// buildLogTargets expands the pod and container dimensions into the (pod,
-// container) product, capped to limit targets overall. An empty container list
-// yields one target per pod (the kubelet default container). It reports whether
-// the cap dropped any target.
-func buildLogTargets(pods, containers []string, limit int) ([]logTarget, bool) {
-	var targets []logTarget
-	for _, p := range pods {
-		if len(containers) == 0 {
-			if len(targets) == limit {
-				return targets, true
-			}
-			targets = append(targets, logTarget{pod: p})
-			continue
-		}
-		for _, c := range containers {
-			if len(targets) == limit {
-				return targets, true
-			}
-			targets = append(targets, logTarget{pod: p, container: c})
-		}
-	}
-	return targets, false
 }
