+++ internal/web/workload_logs_test.go
@@ -8,7 +8,6 @@ import (
 	"errors"
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
 
 	. "github.com/onsi/gomega"
@@ -20,303 +19,9 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
 	fluxcdv1 "github.com/controlplaneio-fluxcd/flux-operator/api/v1"
+	"github.com/controlplaneio-fluxcd/flux-operator/internal/podlogs"
 	"github.com/controlplaneio-fluxcd/flux-operator/internal/web/user"
 )
-
-func TestTrimPartialLogLine(t *testing.T) {
-	tests := []struct {
-		name string
-		in   string
-		want string
-	}{
-		{name: "empty", in: "", want: ""},
-		{name: "newline terminated", in: "line one\nline two\n", want: "line one\nline two\n"},
-		{name: "drops partial trailing line", in: "line one\nline two\npar", want: "line one\nline two\n"},
-		{name: "single partial line kept", in: "partial", want: "partial"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-			g.Expect(trimPartialLogLine(tt.in)).To(Equal(tt.want))
-		})
-	}
-}
-
-func TestTrimPartialFirstLine(t *testing.T) {
-	tests := []struct {
-		name string
-		in   string
-		want string
-	}{
-		{name: "empty", in: "", want: ""},
-		{name: "drops partial leading fragment", in: "tial\nline two\nline three\n", want: "line two\nline three\n"},
-		{name: "single line with no newline kept", in: "only-line", want: "only-line"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-			g.Expect(trimPartialFirstLine(tt.in)).To(Equal(tt.want))
-		})
-	}
-}
-
-func TestTailLogBytes(t *testing.T) {
-	tests := []struct {
-		name             string
-		in               string
-		limit            int
-		want             string
-		wantPartialFirst bool
-	}{
-		{name: "under limit returns all", in: "line1\nline2\n", limit: 100, want: "line1\nline2\n", wantPartialFirst: false},
-		{name: "exact limit not truncated", in: "abcd", limit: 4, want: "abcd", wantPartialFirst: false},
-		// The cut lands exactly after "line1\n", so the first retained line is
-		// complete and must NOT be reported as partial (or the caller drops it).
-		{name: "over limit cut on line boundary keeps complete first line", in: "line1\nline2\nline3\n", limit: 12, want: "line2\nline3\n", wantPartialFirst: false},
-		{name: "over limit cutting mid-line", in: "line1\nline2\n", limit: 8, want: "1\nline2\n", wantPartialFirst: true},
-		{name: "empty stream", in: "", limit: 16, want: "", wantPartialFirst: false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-			// strings.Reader hands out small reads, exercising the chunked loop.
-			got, partialFirst, err := tailLogBytes(strings.NewReader(tt.in), tt.limit)
-			g.Expect(err).NotTo(HaveOccurred())
-			g.Expect(string(got)).To(Equal(tt.want))
-			g.Expect(partialFirst).To(Equal(tt.wantPartialFirst))
-		})
-	}
-}
-
-func TestParseLogEntries(t *testing.T) {
-	t.Run("groups continuation lines with their timestamped entry and stamps the pod and container", func(t *testing.T) {
-		g := NewWithT(t)
-
-		blob := "2026-01-01T00:00:00Z panic: boom\ngoroutine 1 [running]:\nmain.main()\n" +
-			"2026-01-01T00:00:01Z next line\n"
-		entries := parseLogEntries(blob, "pod-a", "app")
-
-		g.Expect(entries).To(HaveLen(2))
-		// The two non-timestamped lines stay attached to the first entry.
-		g.Expect(entries[0].text).To(Equal("2026-01-01T00:00:00Z panic: boom\ngoroutine 1 [running]:\nmain.main()"))
-		g.Expect(entries[0].pod).To(Equal("pod-a"))
-		g.Expect(entries[0].container).To(Equal("app"))
-		g.Expect(entries[1].text).To(Equal("2026-01-01T00:00:01Z next line"))
-		g.Expect(entries[1].pod).To(Equal("pod-a"))
-		g.Expect(entries[1].container).To(Equal("app"))
-	})
-
-	t.Run("a leading continuation with no preceding entry becomes its own entry, stamped", func(t *testing.T) {
-		g := NewWithT(t)
-
-		entries := parseLogEntries("orphan continuation\n2026-01-01T00:00:00Z line\n", "pod-a", "app")
-		g.Expect(entries).To(HaveLen(2))
-		g.Expect(entries[0].ts.IsZero()).To(BeTrue())
-		g.Expect(entries[0].text).To(Equal("orphan continuation"))
-		g.Expect(entries[0].container).To(Equal("app"))
-		g.Expect(entries[1].text).To(Equal("2026-01-01T00:00:00Z line"))
-		g.Expect(entries[1].container).To(Equal("app"))
-	})
-
-	t.Run("empty payload yields no entries", func(t *testing.T) {
-		g := NewWithT(t)
-		g.Expect(parseLogEntries("", "pod-a", "app")).To(BeEmpty())
-	})
-}
-
-func TestMergeLogStreams(t *testing.T) {
-	t.Run("interleaves two streams in chronological order untagged", func(t *testing.T) {
-		g := NewWithT(t)
-
-		app := logStream{pod: "pod-a", blob: "2026-01-01T00:00:00Z app a\n2026-01-01T00:00:02Z app b\n"}
-		sidecar := logStream{pod: "pod-a", blob: "2026-01-01T00:00:01Z side a\n2026-01-01T00:00:03Z side b\n"}
-
-		got := mergeLogStreams([]logStream{app, sidecar}, 0, false, false)
-		g.Expect(got).To(Equal(
-			"2026-01-01T00:00:00Z app a\n" +
-				"2026-01-01T00:00:01Z side a\n" +
-				"2026-01-01T00:00:02Z app b\n" +
-				"2026-01-01T00:00:03Z side b\n"))
-	})
-
-	t.Run("orders fractional timestamps numerically, not lexically", func(t *testing.T) {
-		g := NewWithT(t)
-
-		// "0.12" is numerically after "0.1" but sorts before it lexically; the
-		// merge must parse the timestamps to get this right.
-		a := logStream{pod: "pod-a", blob: "2026-01-01T00:00:00.1Z first\n"}
-		b := logStream{pod: "pod-a", blob: "2026-01-01T00:00:00.12Z second\n"}
-
-		got := mergeLogStreams([]logStream{b, a}, 0, false, false)
-		g.Expect(got).To(Equal("2026-01-01T00:00:00.1Z first\n2026-01-01T00:00:00.12Z second\n"))
-	})
-
-	t.Run("keeps a multi-line entry attached after sorting", func(t *testing.T) {
-		g := NewWithT(t)
-
-		app := logStream{pod: "pod-a", blob: "2026-01-01T00:00:00Z panic\nstack frame\n"}
-		sidecar := logStream{pod: "pod-a", blob: "2026-01-01T00:00:01Z side\n"}
-
-		got := mergeLogStreams([]logStream{app, sidecar}, 0, false, false)
-		g.Expect(got).To(Equal("2026-01-01T00:00:00Z panic\nstack frame\n2026-01-01T00:00:01Z side\n"))
-	})
-
-	t.Run("caps to the newest tailLines entries across all streams", func(t *testing.T) {
-		g := NewWithT(t)
-
-		app := logStream{pod: "pod-a", blob: "2026-01-01T00:00:00Z app a\n2026-01-01T00:00:02Z app b\n"}
-		sidecar := logStream{pod: "pod-a", blob: "2026-01-01T00:00:01Z side a\n2026-01-01T00:00:03Z side b\n"}
-
-		got := mergeLogStreams([]logStream{app, sidecar}, 2, false, false)
-		g.Expect(got).To(Equal("2026-01-01T00:00:02Z app b\n2026-01-01T00:00:03Z side b\n"))
-	})
-
-	t.Run("empty streams yield an empty payload", func(t *testing.T) {
-		g := NewWithT(t)
-		g.Expect(mergeLogStreams(nil, 100, false, false)).To(BeEmpty())
-		g.Expect(mergeLogStreams([]logStream{{pod: "pod-a"}, {pod: "pod-b"}}, 100, false, false)).To(BeEmpty())
-	})
-
-	t.Run("tagged mode prefixes each timestamped line with its pod, interleaved", func(t *testing.T) {
-		g := NewWithT(t)
-
-		a := logStream{pod: "frontend-abc-gqh2x", blob: "2026-01-01T00:00:00Z a one\n2026-01-01T00:00:02Z a two\n"}
-		b := logStream{pod: "frontend-abc-p8x2k", blob: "2026-01-01T00:00:01Z b one\n"}
-
-		got := mergeLogStreams([]logStream{a, b}, 0, true, false)
-		g.Expect(got).To(Equal(
-			"frontend-abc-gqh2x 2026-01-01T00:00:00Z a one\n" +
-				"frontend-abc-p8x2k 2026-01-01T00:00:01Z b one\n" +
-				"frontend-abc-gqh2x 2026-01-01T00:00:02Z a two\n"))
-	})
-
-	t.Run("tagged mode tags only timestamped lines, not continuation or orphan lines", func(t *testing.T) {
-		g := NewWithT(t)
-
-		// A multi-line entry: only its timestamped first line is tagged; the
-		// stack-trace continuation stays untagged so every tagged line is
-		// uniformly "<pod> <timestamp> ...".
-		a := logStream{pod: "pod-a", blob: "2026-01-01T00:00:01Z panic\nstack frame\n"}
-		// A leading orphan continuation (zero ts) is left untagged too.
-		b := logStream{pod: "pod-b", blob: "orphan line\n"}
-
-		got := mergeLogStreams([]logStream{a, b}, 0, true, false)
-		g.Expect(got).To(Equal(
-			"orphan line\n" +
-				"pod-a 2026-01-01T00:00:01Z panic\nstack frame\n"))
-	})
-
-	t.Run("container-tagged mode prefixes each timestamped line with its container", func(t *testing.T) {
-		g := NewWithT(t)
-
-		// Single-pod all-containers view: tag the container, not the pod, so the
-		// client can scope folding per container.
-		app := logStream{pod: "pod-a", container: "app", blob: "2026-01-01T00:00:00Z a one\n2026-01-01T00:00:02Z a two\n"}
-		side := logStream{pod: "pod-a", container: "envoy", blob: "2026-01-01T00:00:01Z s one\n"}
-
-		got := mergeLogStreams([]logStream{app, side}, 0, false, true)
-		g.Expect(got).To(Equal(
-			"app 2026-01-01T00:00:00Z a one\n" +
-				"envoy 2026-01-01T00:00:01Z s one\n" +
-				"app 2026-01-01T00:00:02Z a two\n"))
-	})
-
-	t.Run("combined mode prefixes pod then container, in order", func(t *testing.T) {
-		g := NewWithT(t)
-
-		// All-pods and all-containers: both dimensions fan out, so each timestamped
-		// line carries "<pod> <container> <ts> <msg>"; orphans stay untagged.
-		a := logStream{pod: "p1", container: "app", blob: "2026-01-01T00:00:00Z a\nframe\n"}
-		b := logStream{pod: "p2", container: "side", blob: "2026-01-01T00:00:01Z b\n"}
-
-		got := mergeLogStreams([]logStream{a, b}, 0, true, true)
-		g.Expect(got).To(Equal(
-			"p1 app 2026-01-01T00:00:00Z a\nframe\n" +
-				"p2 side 2026-01-01T00:00:01Z b\n"))
-	})
-}
-
-func TestCapLogBytes(t *testing.T) {
-	tests := []struct {
-		name  string
-		in    string
-		limit int
-		want  string
-	}{
-		{name: "under limit returns all", in: "line1\nline2\n", limit: 100, want: "line1\nline2\n"},
-		{name: "trims partial leading line on a mid-line cut", in: "line1\nline2\nline3\n", limit: 13, want: "line2\nline3\n"},
-		// The cut lands exactly after "line1\n", so the window already starts with
-		// a complete line and the leading line must NOT be dropped.
-		{name: "keeps complete first line when cut lands on a boundary", in: "line1\nline2\nline3\n", limit: 12, want: "line2\nline3\n"},
-		{name: "no newline in window keeps tail bytes", in: "abcdef", limit: 3, want: "def"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-			g.Expect(capLogBytes(tt.in, tt.limit)).To(Equal(tt.want))
-		})
-	}
-}
-
-func TestDedupeNames(t *testing.T) {
-	tests := []struct {
-		name          string
-		in            []string
-		limit         int
-		want          []string
-		wantTruncated bool
-	}{
-		{name: "nil stays nil", in: nil, limit: 8, want: nil},
-		{name: "preserves order without duplicates", in: []string{"app", "sidecar"}, limit: 8, want: []string{"app", "sidecar"}},
-		{name: "drops duplicates keeping first occurrence", in: []string{"app", "app", "sidecar", "app"}, limit: 8, want: []string{"app", "sidecar"}},
-		// De-duplication that leaves the result within the limit is not truncation.
-		{name: "dedup within limit is not truncation", in: []string{"a", "a", "b", "b"}, limit: 2, want: []string{"a", "b"}},
-		{name: "caps to the limit and reports truncation", in: []string{"a", "b", "c", "d"}, limit: 2, want: []string{"a", "b"}, wantTruncated: true},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-			got, truncated := dedupeNames(tt.in, tt.limit)
-			g.Expect(got).To(Equal(tt.want))
-			g.Expect(truncated).To(Equal(tt.wantTruncated))
-		})
-	}
-}
-
-func TestBuildLogTargets(t *testing.T) {
-	t.Run("product of pods and containers", func(t *testing.T) {
-		g := NewWithT(t)
-		targets, truncated := buildLogTargets([]string{"p1", "p2"}, []string{"app", "side"}, 64)
-		g.Expect(truncated).To(BeFalse())
-		g.Expect(targets).To(Equal([]logTarget{
-			{pod: "p1", container: "app"}, {pod: "p1", container: "side"},
-			{pod: "p2", container: "app"}, {pod: "p2", container: "side"},
-		}))
-	})
-
-	t.Run("empty containers yield one default-container target per pod", func(t *testing.T) {
-		g := NewWithT(t)
-		targets, truncated := buildLogTargets([]string{"p1", "p2"}, nil, 64)
-		g.Expect(truncated).To(BeFalse())
-		g.Expect(targets).To(Equal([]logTarget{{pod: "p1"}, {pod: "p2"}}))
-	})
-
-	t.Run("caps to the stream limit and reports truncation", func(t *testing.T) {
-		g := NewWithT(t)
-		targets, truncated := buildLogTargets([]string{"p1", "p2", "p3"}, []string{"app", "side"}, 3)
-		g.Expect(truncated).To(BeTrue())
-		g.Expect(targets).To(HaveLen(3))
-		g.Expect(targets).To(Equal([]logTarget{
-			{pod: "p1", container: "app"}, {pod: "p1", container: "side"},
-			{pod: "p2", container: "app"},
-		}))
-	})
-}
 
 func TestCollectLogStreams(t *testing.T) {
 	errWaiting := errors.New("waiting to start")
@@ -324,9 +29,9 @@ func TestCollectLogStreams(t *testing.T) {
 
 	t.Run("all succeed across two pods", func(t *testing.T) {
 		g := NewWithT(t)
-		targets := []logTarget{{pod: "p1", container: "app"}, {pod: "p2", container: "app"}}
+		targets := []podlogs.LogTarget{{Pod: "p1", Container: "app"}, {Pod: "p2", Container: "app"}}
 		out := collectLogStreams(targets, []string{"x", "y"}, []error{nil, nil})
-		g.Expect(out.streams).To(Equal([]logStream{{pod: "p1", container: "app", blob: "x"}, {pod: "p2", container: "app", blob: "y"}}))
+		g.Expect(out.streams).To(Equal([]podlogs.LogStream{{Pod: "p1", Container: "app", Blob: "x"}, {Pod: "p2", Container: "app", Blob: "y"}}))
 		g.Expect(out.streamedSet).To(HaveLen(2))
 		g.Expect(out.forbidden).To(Equal(0))
 		g.Expect(out.firstErr).NotTo(HaveOccurred())
@@ -334,9 +39,9 @@ func TestCollectLogStreams(t *testing.T) {
 
 	t.Run("partial failure skips the failed target and counts forbidden", func(t *testing.T) {
 		g := NewWithT(t)
-		targets := []logTarget{{pod: "p1", container: "app"}, {pod: "p2", container: "app"}}
+		targets := []podlogs.LogTarget{{Pod: "p1", Container: "app"}, {Pod: "p2", Container: "app"}}
 		out := collectLogStreams(targets, []string{"x", ""}, []error{nil, forbidden})
-		g.Expect(out.streams).To(Equal([]logStream{{pod: "p1", container: "app", blob: "x"}}))
+		g.Expect(out.streams).To(Equal([]podlogs.LogStream{{Pod: "p1", Container: "app", Blob: "x"}}))
 		g.Expect(out.streamedSet).To(HaveLen(1))
 		g.Expect(out.forbidden).To(Equal(1))
 		g.Expect(out.firstErr).To(MatchError(forbidden))
@@ -344,7 +49,7 @@ func TestCollectLogStreams(t *testing.T) {
 
 	t.Run("all fail returns no streams and the first error", func(t *testing.T) {
 		g := NewWithT(t)
-		targets := []logTarget{{pod: "p1", container: "app"}, {pod: "p2", container: "app"}}
+		targets := []podlogs.LogTarget{{Pod: "p1", Container: "app"}, {Pod: "p2", Container: "app"}}
 		out := collectLogStreams(targets, []string{"", ""}, []error{errWaiting, forbidden})
 		g.Expect(out.streams).To(BeEmpty())
 		g.Expect(out.streamedSet).To(BeEmpty())
@@ -355,7 +60,7 @@ func TestCollectLogStreams(t *testing.T) {
 	t.Run("forbidden is counted per pod, not per container target", func(t *testing.T) {
 		g := NewWithT(t)
 		// One forbidden pod with two containers must count once, not twice.
-		targets := []logTarget{{pod: "p1", container: "app"}, {pod: "p1", container: "side"}}
+		targets := []podlogs.LogTarget{{Pod: "p1", Container: "app"}, {Pod: "p1", Container: "side"}}
 		out := collectLogStreams(targets, []string{"", ""}, []error{forbidden, forbidden})
 		g.Expect(out.forbidden).To(Equal(1))
 	})
@@ -368,7 +73,7 @@ func TestBuildWorkloadLogsResponse(t *testing.T) {
 		g := NewWithT(t)
 		// Two pods requested, only p1 streams; p2 is forbidden. The response must
 		// report 200-style partial coverage with the per-pod counts the UI shows.
-		targets := []logTarget{{pod: "p1", container: "app"}, {pod: "p2", container: "app"}}
+		targets := []podlogs.LogTarget{{Pod: "p1", Container: "app"}, {Pod: "p2", Container: "app"}}
 		fanOut := collectLogStreams(targets, []string{"2026-01-01T00:00:00Z hi\n", ""}, []error{nil, forbidden})
 		resp := buildWorkloadLogsResponse([]string{"p1", "p2"}, []string{"app"}, fanOut, 2, 1000, true, false, false)
 
@@ -384,7 +89,7 @@ func TestBuildWorkloadLogsResponse(t *testing.T) {
 
 	t.Run("full coverage is not partial", func(t *testing.T) {
 		g := NewWithT(t)
-		targets := []logTarget{{pod: "p1", container: "app"}, {pod: "p2", container: "app"}}
+		targets := []podlogs.LogTarget{{Pod: "p1", Container: "app"}, {Pod: "p2", Container: "app"}}
 		fanOut := collectLogStreams(targets, []string{"2026-01-01T00:00:00Z a\n", "2026-01-01T00:00:01Z b\n"}, []error{nil, nil})
 		resp := buildWorkloadLogsResponse([]string{"p1", "p2"}, []string{"app"}, fanOut, 2, 1000, true, false, false)
 
@@ -397,7 +102,7 @@ func TestBuildWorkloadLogsResponse(t *testing.T) {
 		g := NewWithT(t)
 		// total counts requested pods (1) and that pod streamed, but the fan-out
 		// was capped, so the response must still flag the result as partial.
-		targets := []logTarget{{pod: "p1", container: "app"}}
+		targets := []podlogs.LogTarget{{Pod: "p1", Container: "app"}}
 		fanOut := collectLogStreams(targets, []string{"2026-01-01T00:00:00Z x\n"}, []error{nil})
 		resp := buildWorkloadLogsResponse([]string{"p1"}, []string{"app"}, fanOut, 1, 1000, true, false, true)
 
@@ -410,7 +115,7 @@ func TestBuildWorkloadLogsResponse(t *testing.T) {
 		g := NewWithT(t)
 		// One pod, two containers: ContainerTagged is set and each line is prefixed
 		// with its container so the client scopes folding per container.
-		targets := []logTarget{{pod: "p1", container: "app"}, {pod: "p1", container: "envoy"}}
+		targets := []podlogs.LogTarget{{Pod: "p1", Container: "app"}, {Pod: "p1", Container: "envoy"}}
 		fanOut := collectLogStreams(targets, []string{"2026-01-01T00:00:00Z a\n", "2026-01-01T00:00:01Z e\n"}, []error{nil, nil})
 		resp := buildWorkloadLogsResponse([]string{"p1"}, []string{"app", "envoy"}, fanOut, 1, 1000, false, true, false)
